pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "huawei-pdf-reader"
version = "0.1.0"
description = "华为平板PDF阅读器 - 支持PDF/Word阅读、手写笔注释、翻译和繁简转换"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Developer", email = "dev@example.com"}
]
keywords = ["pdf", "reader", "huawei", "tablet", "annotation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "kivy>=2.2.0",
    "PyMuPDF>=1.23.0",
    "python-docx>=1.1.0",
    "opencc-python-reimplemented>=0.1.7",
    "paddleocr>=2.7.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.92.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]

[project.scripts]
huawei-pdf-reader = "huawei_pdf_reader.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.hypothesis]
max_examples = 100
verbosity = "normal"

[tool.coverage.run]
source = ["src/huawei_pdf_reader"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
